Move kotlin-dsl specific build logic to buildSrc plugins
