+++ subprojects/kotlin-compiler-embeddable/kotlin-compiler-embeddable.gradle.kts
@@ -1,9 +1,6 @@
-import gradlebuild.basics.BuildEnvironment
-import gradlebuild.kotlindsl.tasks.CheckKotlinCompilerEmbeddableDependencies
-import gradlebuild.kotlindsl.tasks.PatchKotlinCompilerEmbeddable
-
 plugins {
     id("gradlebuild.distribution.implementation-kotlin")
+    id("gradlebuild.kotlin-dsl-compiler-embeddable")
 }
 
 description = "Kotlin Compiler Embeddable - patched for Gradle"
@@ -17,51 +14,4 @@ dependencies {
     api(libs.futureKotlin("daemon-embeddable"))
 
     runtimeOnly(libs.trove4j)
-}
-
-val kotlinCompilerEmbeddable by configurations.creating
-
-dependencies {
-    kotlinCompilerEmbeddable(project(":distributionsDependencies"))
-    kotlinCompilerEmbeddable(libs.futureKotlin("compiler-embeddable"))
-}
-
-tasks {
-
-    val checkKotlinCompilerEmbeddableDependencies by registering(CheckKotlinCompilerEmbeddableDependencies::class) {
-        current.from(configurations.runtimeClasspath)
-        expected.from(kotlinCompilerEmbeddable)
-    }
-
-    val patchKotlinCompilerEmbeddable by registering(PatchKotlinCompilerEmbeddable::class) {
-        dependsOn(checkKotlinCompilerEmbeddableDependencies)
-        excludes.set(listOf(
-            "META-INF/services/javax.annotation.processing.Processor",
-            "META-INF/native/**/*jansi.*"
-        ))
-        originalFiles.from(kotlinCompilerEmbeddable)
-        dependencies.from(configurations.detachedConfiguration(
-            project.dependencies.project(":distributionsDependencies"),
-            project.dependencies.create(libs.jansi)
-        ))
-        dependenciesIncludes.set(mapOf(
-            "jansi-" to listOf("META-INF/native/**", "org/fusesource/jansi/internal/CLibrary*.class")
-        ))
-        additionalRootFiles.from(classpathManifest)
-
-        outputFile.set(jar.get().archiveFile)
-
-        outputs.doNotCacheIfSlowInternetConnection()
-    }
-
-    jar {
-        dependsOn(patchKotlinCompilerEmbeddable)
-        actions.clear()
-    }
-}
-
-fun TaskOutputs.doNotCacheIfSlowInternetConnection() {
-    doNotCacheIf("Slow internet connection") {
-        BuildEnvironment.isSlowInternetConnection
-    }
 }
